Stored prediction results in DB

+++ predictapp/models.py
@@ -5,11 +5,11 @@ from django.db import models
 class Record(models.Model):
     gender = models.CharField(max_length=20)
     grade = models.IntegerField()
-    openness = models.FloatField()
-    agreeableness = models.FloatField()
-    emotions = models.FloatField()
-    conscientiousness = models.FloatField()
-    extraversion = models.FloatField()
+    openness = models.IntegerField()
+    agreeableness = models.IntegerField()
+    emotions = models.IntegerField()
+    conscientiousness = models.IntegerField()
+    extraversion = models.IntegerField()
     course = models.CharField(max_length=30)
 
     def __str__(self):
